+++ R/EpiTracer-package.R
@@ -17,5 +17,5 @@ utils::globalVariables(c(
   "strands", "gene", "gene_name", "pos_label", "min_coord", "max_coord",
   "Polyploidy", "y", "label", "pos", "gieStain", "color",
   "x", "xend", "yend", "lwd", "colour", "grp",
-  "gx", "gx_start", "gx_end", "amp_bp"
+  "gx", "gx_start", "gx_end", "amp_bp", "gxs", "gxe"
 ))